+++ R/controls.R
@@ -11,7 +11,8 @@
 #' - `linestring`: draw linestrings by clicking each vertex
 #' - `polygon`: draw polygons by clicking each vertex
 #' - `lasso`: freehand polygon draw by click-dragging
-#' @param features <`sf` | `sfc`> Features with which to initialise the editor
+#' @param features <[`wk-geometry`]> Features with which to initialise the editor.
+#' Requires CRS [EPSG:4326](http://epsg.io/4326).
 #' @export
 editor_options <- function(mode = cur_value(), features = cur_value()) {
   tidyassert::assert(
@@ -22,13 +23,16 @@ editor_options <- function(mode = cur_value(), features = cur_value()) {
   )
 
   tidyassert::assert(
-    is.null(features) ||
-      is_cur_value(features) ||
-      (is_sf(features) || is_sfc(features)) && is_wgs84(features),
+    is.null(features) || is_cur_value(features) ||
+      wk::is_handleable(features) && is_wgs84(features),
     error_message = c(
-      "x" = "{.arg features} must be a {.emph WGS84} {.cls sf/sfc}"
+      "x" = "{.arg features} must be a {.emph WGS84} {.cls wk-geometry}"
     )
   )
+
+  if (inherits(features, "data.frame")) {
+    features <- purrr::detect(features, wk::is_handleable)
+  }
 
   structure(
     list(
@@ -51,6 +55,7 @@ editor_modes <- function() {
 is_editor_options <- function(object) inherits(object, "editor_options")
 
 as_editor_options <- function(object) UseMethod("as_editor_options")
-as_editor_options.default <- function(object) object
+as_editor_options.editor_options <- function(object) object
+as_editor_options.cur_value <- function(object) object
 as_editor_options.NULL <- function(object) NULL
 as_editor_options.logical <- function(object) if (isTRUE(object)) editor_options() else NULL